refactor: join "entity" and "repository" package into package "repositories"
