--- entity/lib/entity.dart
@@ -1,9 +1,0 @@
-export 'id.dart';
-
-class Entity {
-  const Entity();
-}
-
-class Nullable {
-  const Nullable();
-}
