--- entity/pubspec.yaml
@@ -1,10 +1,0 @@
-name: entity
-version: 1.0.0
-
-environment:
-  sdk: ">=2.0.0 <3.0.0"
-
-dependencies:
-  
-dev_dependencies:
-  test: ^1.0.0
