--- entity_generator/.dart_tool/pub/bin/sdk-version
@@ -1,1 +1,0 @@
-2.2.0
