+++ entity_generator/.packages
@@ -1,4 +1,4 @@
-# Generated by pub on 2019-08-20 21:55:43.890525.
+# Generated by pub on 2019-08-20 23:15:40.755267.
 analyzer:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/analyzer-0.32.6/lib/
 args:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/args-1.5.2/lib/
 async:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/async-2.3.0/lib/
@@ -13,16 +13,17 @@ built_collection:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.da
 built_value:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/built_value-6.7.0/lib/
 charcode:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/charcode-1.1.2/lib/
 code_builder:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/code_builder-3.2.0/lib/
-collection:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/collection-1.14.12/lib/
+collection:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/collection-1.14.11/lib/
 convert:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/convert-2.1.1/lib/
 crypto:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/crypto-2.1.2/lib/
 csslib:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/csslib-0.14.6+1/lib/
 dart_style:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/dart_style-1.1.3/lib/
-entity:../entity/lib/
 fixnum:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/fixnum-0.10.9/lib/
+flutter:file:///C:/src/flutter/packages/flutter/lib/
 front_end:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/front_end-0.1.4+2/lib/
 glob:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/glob-1.1.7/lib/
 graphs:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/graphs-0.1.3+1/lib/
+hive:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/hive-0.4.1/lib/
 html:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/html-0.14.0/lib/
 http:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/http-0.12.0+2/lib/
 http_multi_server:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/http_multi_server-2.1.0/lib/
@@ -41,16 +42,23 @@ node_preamble:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartl
 package_config:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/package_config-1.1.0/lib/
 package_resolver:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/package_resolver-1.0.10/lib/
 path:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/path-1.6.4/lib/
+path_provider:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/path_provider-1.2.0/lib/
 pedantic:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/pedantic-1.8.0+1/lib/
 plugin:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/plugin-0.2.0+3/lib/
+pointycastle:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/pointycastle-1.0.1/lib/
 pool:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/pool-1.4.0/lib/
+provider:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/provider-3.1.0/lib/
 pub_semver:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/pub_semver-1.4.2/lib/
 pubspec_parse:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/pubspec_parse-0.1.4/lib/
 quiver:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/quiver-2.0.5/lib/
+repositories:../repositories/lib/
+rxdart:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/rxdart-0.22.1+1/lib/
+shared_preferences:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/shared_preferences-0.5.3+4/lib/
 shelf:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/shelf-0.7.5/lib/
 shelf_packages_handler:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/shelf_packages_handler-1.0.4/lib/
 shelf_static:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/shelf_static-0.2.8/lib/
 shelf_web_socket:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/shelf_web_socket-0.2.3/lib/
+sky_engine:file:///C:/src/flutter/bin/cache/pkg/sky_engine/lib/
 source_gen:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/source_gen-0.9.1+3/lib/
 source_map_stack_trace:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/source_map_stack_trace-1.1.5/lib/
 source_maps:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/source_maps-0.10.8/lib/
@@ -65,6 +73,7 @@ test_api:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.o
 test_core:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/test_core-0.2.1+1/lib/
 timing:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/timing-0.1.1+2/lib/
 typed_data:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/typed_data-1.1.6/lib/
+vector_math:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/vector_math-2.0.8/lib/
 vm_service_client:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/vm_service_client-0.2.6+3/lib/
 watcher:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/watcher-0.9.7+12/lib/
 web_socket_channel:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/web_socket_channel-1.0.15/lib/
