+++ entity_generator/lib/builder.dart
@@ -1,6 +1,6 @@
-import 'package:analyzer/analyzer.dart';
+import 'package:analyzer/dart/element/element.dart';
 import 'package:build/build.dart';
-import 'package:entity/entity.dart';
+import 'package:repositories/repositories.dart';
 import 'package:source_gen/source_gen.dart';
 
 class EntityGenerator extends GeneratorForAnnotation<Entity> {
