+++ entity_generator/pubspec.yaml
@@ -2,13 +2,14 @@ name: entity_generator
 version: 1.0.0
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.2.0 <3.0.0"
 
 dependencies:
+  repositories:
+    path: ../repositories
+
   build: '>=0.12.0 <2.0.0'
   source_gen: ^0.9.0
-  entity:
-    path: ../entity/
   
 dev_dependencies:
   build_test: ^0.10.0
