+++ repositories/.flutter-plugins
@@ -1,0 +1,2 @@
+path_provider=C:\\Users\\Marcel\\AppData\\Roaming\\Pub\\Cache\\hosted\\pub.dartlang.org\\path_provider-1.2.0\\
+shared_preferences=C:\\Users\\Marcel\\AppData\\Roaming\\Pub\\Cache\\hosted\\pub.dartlang.org\\shared_preferences-0.5.3+4\\
