+++ repositories/.packages
@@ -1,0 +1,59 @@
+# Generated by pub on 2019-08-20 23:10:26.914489.
+analyzer:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/analyzer-0.38.0/lib/
+args:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/args-1.5.2/lib/
+async:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/async-2.3.0/lib/
+boolean_selector:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/boolean_selector-1.0.5/lib/
+charcode:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/charcode-1.1.2/lib/
+collection:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/collection-1.14.11/lib/
+convert:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/convert-2.1.1/lib/
+crypto:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/crypto-2.1.2/lib/
+csslib:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/csslib-0.16.1/lib/
+flutter:file:///C:/src/flutter/packages/flutter/lib/
+front_end:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/front_end-0.1.22/lib/
+glob:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/glob-1.1.7/lib/
+hive:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/hive-0.4.1/lib/
+html:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/html-0.14.0+2/lib/
+http:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/http-0.12.0+2/lib/
+http_multi_server:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/http_multi_server-2.1.0/lib/
+http_parser:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/http_parser-3.1.3/lib/
+io:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/io-0.3.3/lib/
+js:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/js-0.6.1+1/lib/
+kernel:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/kernel-0.3.22/lib/
+matcher:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/matcher-0.12.5/lib/
+meta:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/meta-1.1.7/lib/
+mime:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/mime-0.9.6+3/lib/
+multi_server_socket:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/multi_server_socket-1.0.2/lib/
+node_preamble:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/node_preamble-1.4.6/lib/
+package_config:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/package_config-1.1.0/lib/
+package_resolver:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/package_resolver-1.0.10/lib/
+path:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/path-1.6.4/lib/
+path_provider:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/path_provider-1.2.0/lib/
+pedantic:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/pedantic-1.8.0+1/lib/
+pointycastle:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/pointycastle-1.0.1/lib/
+pool:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/pool-1.4.0/lib/
+provider:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/provider-3.1.0/lib/
+pub_semver:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/pub_semver-1.4.2/lib/
+rxdart:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/rxdart-0.22.1+1/lib/
+shared_preferences:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/shared_preferences-0.5.3+4/lib/
+shelf:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/shelf-0.7.5/lib/
+shelf_packages_handler:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/shelf_packages_handler-1.0.4/lib/
+shelf_static:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/shelf_static-0.2.8/lib/
+shelf_web_socket:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/shelf_web_socket-0.2.3/lib/
+sky_engine:file:///C:/src/flutter/bin/cache/pkg/sky_engine/lib/
+source_map_stack_trace:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/source_map_stack_trace-1.1.5/lib/
+source_maps:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/source_maps-0.10.8/lib/
+source_span:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/source_span-1.5.5/lib/
+stack_trace:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/stack_trace-1.9.3/lib/
+stream_channel:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/stream_channel-2.0.0/lib/
+string_scanner:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/string_scanner-1.0.5/lib/
+term_glyph:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/term_glyph-1.1.0/lib/
+test:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/test-1.6.7/lib/
+test_api:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/test_api-0.2.6/lib/
+test_core:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/test_core-0.2.8/lib/
+typed_data:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/typed_data-1.1.6/lib/
+vector_math:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/vector_math-2.0.8/lib/
+vm_service:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/vm_service-1.1.0/lib/
+watcher:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/watcher-0.9.7+12/lib/
+web_socket_channel:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/web_socket_channel-1.0.15/lib/
+yaml:file:///C:/Users/Marcel/AppData/Roaming/Pub/Cache/hosted/pub.dartlang.org/yaml-2.1.16/lib/
+repositories:lib/
