+++ repositories/entity.dart
@@ -1,0 +1,7 @@
+class Entity {
+  const Entity();
+}
+
+class Nullable {
+  const Nullable();
+}
