+++ repositories/pubspec.yaml
@@ -1,0 +1,18 @@
+name: repositories
+version: 1.0.0
+
+environment:
+  sdk: ">=2.2.0 <3.0.0"
+
+dependencies:
+  flutter:
+    sdk: flutter
+
+  hive: ^0.4.1
+  path_provider: ^1.1.0
+  provider: ^3.0.0
+  rxdart: ^0.22.0
+  shared_preferences: ^0.5.3
+
+dev_dependencies:
+  test: ^1.0.0
